requirements.txt
anthropic>=0.39.0
httpx>=0.27.0
pydantic>=2.0.0
apscheduler>=3.10.0
click>=8.0.0
python-dotenv>=1.0.0
